added mysql scale safety check with table scan checker
